refactor(#72): move hooks and components into directories
