+++ src/index.ts
@@ -1,3 +1,3 @@
-export { GoogleAnalytics } from "./GoogleAnalytics";
-export { usePagesViews } from "./usePagesViews";
+export { GoogleAnalytics } from "./components";
+export { usePagesViews } from "./hooks";
 export { pageView, event } from "./interactions";
